Refactored AboutSection so that it's rendered version is it's own component, better for useEffect purposes
