+++ src/components/AboutSection/AboutSection.js
@@ -1,19 +1,12 @@
-import { useState, useEffect } from "react"
+import {useState} from "react"
 import "./AboutSection.scss"
 import SectionTitle from "../../components/SectionTitle/SectionTitle"
+import RenderedAbout from "../RenderedAbout/RenderedAbout"
 
 // Disable warning for incorrect usage of ${} since I'm intentionally trying to print ${} to the screen.
 /* eslint-disable no-template-curly-in-string */ 
 
   const AboutSection = () => {
-  const timeOutArray = []
-  const intervalArray = []
-  const [titleText, setTitleText] = useState("")
-  const [p1Text, setP1Text] = useState("")
-  const [p2Text, setP2Text] = useState("")
-  const [p3Text, setP3Text] = useState("")
-  const [span1Text, setSpan1Text] = useState("")
-  const [textCursor, setTextCursor] = useState('title')
   const [showText, setShowText] = useState(false)
   const [nameInput, setNameInput] = useState({value: "Keith Ryan O'Rourke", length: 19, style: {width: "19ch"}}) 
   const [developerInput, setDeveloperInput] = useState({value: "full stack", length: 10, style: {width: "10ch"}}) 
@@ -32,67 +25,13 @@ import SectionTitle from "../../components/SectionTitle/SectionTitle"
     }
   }
 
-  /**
-   * 
-   * @param {string} initialState 
-   * @param {string} content 
-   * @param {function} setter 
-   * @param {integer} interval 
-   * @param {string} nextCursor 
-   * @returns {number} time in ms to complete animation
-   */
-  const animateString = (delayTime, content, setter, interval, nextCursor) => {
-    let newString = ''
-    let i = 0
-    console.log(delayTime, content, i, nextCursor, content.length)
-    timeOutArray.push(setTimeout(() => {
-      const animationTimer = setInterval(() => {
-        if(content[i]) {
-          newString += content[i]
-          setter(newString)
-        }
-        i++
-        if(i >=content.length || !content[i]) {
-          clearInterval(animationTimer)
-          setTextCursor(prevState => nextCursor || prevState)
-        }
-      }, interval)
-      intervalArray.push(animationTimer)
-    }, delayTime || 0))
-    return (interval*content.length)
-  }
-
-  const renderText = () => {
-    document.querySelector('.app').scroll(0, 0)
-    const splitName = nameInput.value.split(' ')
-    const middlename = (splitName.length === 3 ? splitName[1] || 'Ryan' : 'Ryan')
-    let delayTime = 0;
-    delayTime += animateString(0, "Hello!", setTitleText, 50, 'p1')
-    delayTime += animateString(delayTime+300, `My name is ${nameInput.value || "(Hm... Seems like someone left a field empty!)"} and I'm a ${developerInput.value || "(Hm... Seems like someone left a field empty!)"} web developer!`, setP1Text, 25, 'p2')
-    delayTime += animateString(delayTime+500, "My favorite thing about web development is the plethora of opportunities to collaborate with others. I'm very passionate about technology education and professional growth!", setP2Text, 25, 'p3')
-    delayTime += animateString(delayTime+500, `By the way: I go by my middlename, ${middlename}! `, setP3Text, 25, 'span1')
-    animateString(delayTime+500, '〜(￣▽￣〜)', setSpan1Text, 25, "")
-  }
-
-  // Cleanup useEffect in case user leaves page before text animations complete
-  useEffect(() => {
-    return () => {
-    for(let i=0; i<timeOutArray.length; i++) {
-      clearTimeout(timeOutArray[i])
-    }
-    for (let i=0; i<intervalArray.length; i++) {
-      clearInterval(intervalArray[i])
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }}, [])
-
   const submitHandler = (e) => {
     e.preventDefault()
     if(e.target.love.value !== 'true') {
       alert('Lies!!')
     } else {
       setShowText(true)
-      renderText()
+      // renderText()
     }
   }
   return (
@@ -149,12 +88,10 @@ import SectionTitle from "../../components/SectionTitle/SectionTitle"
         </div>
       </form>
       : 
-      <div className="about__codebox">
-        <h3 className="about__subtitle">{titleText}{textCursor === 'title' ? <div className="about__text-cursor"></div> : ""}</h3>
-        <p className="about__copy">{p1Text}{textCursor === 'p1' ? <span className="about__text-cursor"></span> : ""}</p>
-        <p className="about__copy">{p2Text}{textCursor === 'p2' ? <span className="about__text-cursor"></span> : ""}</p>
-        <p className="about__copy">{p3Text}{textCursor === 'p3' ? <span className="about__text-cursor"></span> : ""}<span className="about__face">{span1Text}{textCursor === 'span1' ? <span className="about__text-cursor"></span> : ""}</span></p>
-      </div>
+      <RenderedAbout
+        nameInput={nameInput}
+        developerInput={developerInput}
+      />
       }
     </section>
       )
